pyproject.toml
[project]
name = "content-creator-agent"
version = "0.1.0"
description = "AI-powered content creation agent for LinkedIn and X/Twitter"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # === Core Framework ===
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "openai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "celery[redis]>=5.4.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",

    # === Research ===
    "feedparser>=6.0.11",
    "praw>=7.8.0",
    "arxiv>=2.1.3",

    # === Messaging ===
    "slack-bolt>=1.21.0",
    "slack-sdk>=3.33.0",

    # === Social Publishing ===
    "tweepy>=4.14.0",

    # === Vector / Embeddings ===
    "pgvector>=0.3.6",
    "voyageai>=0.3.0",

    # === Utilities ===
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "pytz>=2024.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "respx>=0.22.0",
    "pytest-mock>=3.14.0",
    "factory-boy>=3.3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["agent"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
